+++ folio-spring-cql/src/test/java/org/folio/spring/cql/domain/CapabilitySet.java
@@ -1,0 +1,34 @@
+package org.folio.spring.cql.domain;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import java.util.UUID;
+import lombok.Data;
+import org.hibernate.annotations.JdbcTypeCode;
+import org.hibernate.type.SqlTypes;
+
+@Entity
+@Data
+@Table(name = "capability_set")
+public class CapabilitySet {
+
+  @Id
+  @GeneratedValue
+  private UUID id;
+
+  @Column(name = "name")
+  private String name;
+
+  @Column(name = "resource")
+  private String resource;
+
+  @Enumerated(EnumType.STRING)
+  @JdbcTypeCode(SqlTypes.NAMED_ENUM)
+  @Column(name = "type", columnDefinition = "capability_type")
+  private EntityCapabilityType type;
+}