+++ folio-spring-cql/src/test/java/org/folio/spring/cql/repo/CapabilitySetRepository.java
@@ -1,0 +1,8 @@
+package org.folio.spring.cql.repo;
+
+import java.util.UUID;
+import org.folio.spring.cql.JpaCqlRepository;
+import org.folio.spring.cql.domain.CapabilitySet;
+
+public interface CapabilitySetRepository extends JpaCqlRepository<CapabilitySet, UUID> {
+}